1F1: Add missing typecast to enum.

+++ include/boost/math/special_functions/detail/hypergeometric_1F1_large_abz.hpp
@@ -48,7 +48,7 @@
         {
            typedef typename lanczos::lanczos<T, Policy>::type lanczos_type;
 
-           gamma_cache[cache_size - 1] = boost::math::gamma_p(alpha_poch + cache_size - 1, x, pol);
+           gamma_cache[cache_size - 1] = boost::math::gamma_p(alpha_poch + (int)cache_size - 1, x, pol);
            for (int i = cache_size - 1; i > 0; --i)
            {
               gamma_cache[i - 1] = gamma_cache[i] >= 1 ? 1 : gamma_cache[i] + regularised_gamma_prefix(alpha_poch + i - 1, x, pol, lanczos_type()) / (alpha_poch + i - 1);
